+++ src/main/java/com/cliptripbe/feature/place/application/PlaceTranslationService.java
@@ -7,8 +7,8 @@ import com.cliptripbe.feature.place.domain.entity.PlaceTranslation;
 import com.cliptripbe.feature.place.domain.vo.TranslationInfo;
 import com.cliptripbe.feature.place.infrastructure.PlaceTranslationRepository;
 import com.cliptripbe.feature.user.domain.type.Language;
-import com.cliptripbe.infrastructure.openai.prompt.type.PromptConstants;
-import com.cliptripbe.infrastructure.openai.service.ChatGPTService;
+import com.cliptripbe.infrastructure.adapter.out.openai.prompt.type.PromptConstants;
+import com.cliptripbe.infrastructure.adapter.out.openai.service.ChatGPTService;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
