+++ src/main/java/com/cliptripbe/infrastructure/port/caption/CaptionPort.java
@@ -1,0 +1,9 @@
+package com.cliptripbe.infrastructure.port.caption;
+
+import com.cliptripbe.infrastructure.adapter.out.caption.dto.CaptionResponse;
+
+public interface CaptionPort {
+
+    CaptionResponse getCaptions(String youtubeUrl);
+
+}
